requirements.txt
aiohttp==3.7.4
asgiref==3.7.2
async-timeout==3.0.1
attrs==23.2.0
bson==0.5.10
certifi==2019.9.11
chardet==3.0.4
click==8.0.3
colorama==0.4.6
commonmark==0.9.1
dj-database-url==2.1.0
Django==4.2.11
django-ckeditor==6.7.1
django-jazzmin==2.6.0
django-js-asset==2.2.0
djangorestframework==3.15.1
idna==3.6
multidict==6.0.5
pillow==10.2.0
psycopg2-binary==2.9.9
Pygments==2.17.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
qrcode==7.4.2
rich==10.14.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
websockets==9.1
yarl==1.9.4
